Create WordCountModified.java
Modify the Word Count program so that it only counts “real” words, in particular, strings that start with letters “m”, “n”, “o”, “p”, and “q” (no matter if they are capitalized or not).
Add a custom partitioner to this Word Count program. It should assign words starting with “M” or “m” to reduce task 0, those starting with “N” or “n” to task 1, and so on.